cq_zero(): now always nullify the event, as the name suggests.
We used to not do that for "foreign" events, forcing them to call cq_cancel()
but that was not good -- it was easy to forget, not realizing a particular
call would be "foreign".

So now cq_zero() always clears the event, which means the applicaiton must
take a lock before calling it and before registing the event, to avoid race
conditions.  This also means cq_zero_if_triggered() is no longer needed since
there cannot be any concurrent triggering and zeroing of the event as long as
the lock is held!

There is still a gap though, as mentioned in the wd_expired() routine and also
in the new cq_event_called() routine, but it would require a huge interface
change in all callout event-recording routines.

To protect against this race condition bug, we have now a mandatory warning
issued when they call cq_zero() on a "foreign" event with no locks held.  Just
to pinpoint a possible problem source with this old API.