thread_sig_handle(): use atomic exchange operation to clear pending set.
Since posting of a signal uses a similar atomic exchange to record the pending
signal, it was an oversight when we switched to that atomic exchange to not
also clear the pending set with a similar exchange.  Locking the thread element
was useless to prevent races and could have lead to signal losses.