Use new cq_event() in thread callout events to avoid possible leaks.
If the event is "foreign" to the callout queue, not acknowledging it
would make it leak since the reference count on the event will not be
properly decreased after its invocation.

Fortunately, cq_expire_internal() actually does the ev_free() call for
us but it also loudly warns, so when we can, calling cq_event() will set
things right and avoid the warning.