node_bye_v(): do not call node_remove_v() for G2 nodes.
We already consumed the variable arguments on the stack, we cannot call
node_remove_v() which would attempt to consume others on the stack, not
the ones we want!  This could lead to unpredictable errors, including a
memory access violation.

Clone the error string we already built, and invoke node_remove() safely.