Simplified count of key occurrences in aging table.
The new API that we can now use is made of:

	aging_saw_another_revitalise()
	aging_saw_another()

to record occurrences with or without revitalisation of the key, and

	aging_seen_count()

to request the actual amount of occcurrences registered so far.

This is also slightly more efficient because we no longer need to
blindly allocate a copy of the key from the caller: this is now done
as needed, by the functions, based on a supplied key-cloning routine.