Critical upload HTTP statuses are now fully sent.
During errors, we do not necessarily care whether we can send the whole
HTTP status back.

However, when serving files or actively queueing, not being able to send
the whole HTTP header was a fatal error, yielding to a bad upload service.

We now make sure we are buffering the unsent part of the header and
continue to send the missing parts asynchronously to preserve the
connection with the remote servent when we are expecting follow-up
HTTP requests from them, possibly.