Do not attempt costly operations after a crash / auto-restart.
Things like the SDBM database rebuilding that we normally do during
normal startup can easily take a very long time.  When we are restarting
after a crash, or when the user requests the auto-restart from the
shell, we want to speed-up the application initialization path.

To that end, the crash layer now records whether the application can
determine whether it was restarted.  And that can be later queried to
skip long processing during initializations.