Be sure to use destructive list iteration with tmalloc_free_magazine().
Since we use embedded list pointers in the objects and we are iterating
to free-up the item, we need to remove the item from the list or we are
preventing any further traversal on that list.

This was not a problem under normal operations, but when safety
assertions were turned on, it could create random crashes as some of the
additional assertions are attempting to traverse the list!  When the
freed memory was reused by concurrent threads, boom...