shared_file_path(): must process sf->fi correctly!
We were not checking whether sf->fi was NULL, always returning
sf->file_path.  That is OK for library files, but partial files which we
are sharing need to refer to the fileinfo's path, to be able to be
served correctly in "Seeding" mode, when they have been possibly moved.

This was the root cause of a bug whereby completed files, normally
seeded in "partial file sharing" mode, could in fact never be served.