file_info_retrieve(): must restore proper mtime for seeded files!
Unfortunately, unless the mtime is correct, the upload layer will refuse
to share the file when requested, assuming that since it was modified on
disk, it may not reflect the original file we downloaded and verified.

For now, we use the "TIME" field to persist the last known modification
time, restoring it correctly for seeded files.  We also emit a warning
if that timestamp is not in sync with the one on disk, and reset it.