file_info_retrieve(): check seeded file against SHA1 cache.
This better handles mtime inconsistency when resuming file seeding.

When the modified time in the fileinfo does not match that of the file
on disk, check whether the information on disk matches with the one in
the SHA1 cache.

If the information matches, chances are that the fileinfo database was
wrong and we can keep the file.  Otherwise, stop seeding that file.