Persist current_peermode to restart in same mode after crash.
This matters when they left the configured peermode as "automatic", the
default value, and the crash happens when the node was already running
as an ultrapeer.  We do not want to restart as a leaf node to get promoted
later when we know we were already promoted in the past.