vmm_move(): spurious vmm_free() was corrupting stats.
When we are freeing the manually allocated pages, we must not call
vmm_free(), which will account the pages being freed whereas we have
not accounted them as being allocated yet!

Over time, this was leading to an assertion failure because the amount
of pages being tracked as allocated became negative!