Do not attempt to create VMM magazines when allocating core.
In vmm_alloc_core(), do not call vmm_get_magazine() as that could create
a magazine in the thread if it does not already exists, possibly
attempting to recurse to allocate more core.

Likewise, use the same logic in vmm_move_internal(): regardless of
whether we allocate user or core memory, if there is no depot yet for
the given size, then there is no need to invoke tmalloc_smart() anyway.