Added cstr_{b,f,l}cpy() routine to supplant g_strlcpy().
The motivation is that of the signature: we wish to have the destination
and length parameters following eachother so that we can use constructs
like ARYLEN() when possible, to factorize the variable name.

The variants are provided for convenience: cstr_lcpy() is an emulation
of strlcpy(), with just a different signature. The cstr_bcpy() routine
behaves like cstr_lcpy() but will loudly warn if its destination buffer
is too small.  The cstr_fcpy() routine simply returns a boolean
indicating whether the string fitted (to again, avoid having to spell
out the buffer variable name).