compat_longjmp(): record usage location and trace duplicate use.
In case of duplicate use of a jmp_buf or sigjmp_buf, we now trace where
the corresponding longjmp() or siglongjmp() call was made, along with
the passed value, in an attempt to help understand, post-mortem, why there
was a duplicate use of the saved context.

Removed the check for sigsetjmp() being used after longjmp(): it does
not make sense since the structures are not identical (sigjmp_buf versus
jmp_buf_) so the compiler should catch misuses already.

Removed calls to g_error(), use s_error() from the library.