BG: record whether task is in "user" mode.
This fixes a bug when cancelling a running task that is already outside
its setjmp() protection that identifies the "user" mode, where we are
running user code and can call bg_task_exit().

In particular, it was causing a crash at exit time when we were
attempting to cancel a running task, for instance the TTH checking
task running on the library in the background.