Fix stat() and fstat() on Windows.
We now compute meaningful st_dev and st_ino fields, and supply
synthetized st_blksize and st_blocks fields in the filestat_t
structure.

The st_mode field is also sanitized to drop the trailing bits that
we ignore on Windows (those for group and others).

Finally, when conducting fstat() tests, as long as we are running on
Vista or higher (Vista comes after XP), we specifically check whether
the associated filename is an executable (based on its extension)
to force the 'x' rights.