Semi-reliable UDP TX counts sent/unsent reliable messages.
We now count reliable messages sent/unsent separately.

The amount of classes was also increased from 6 to 10, in order to see
how the frequency of unsent messages increases with the size of the
message to transmit.

It also helps see how the size of the messages to send distributes among
these classes, those with 10+ fragments being larger than 4KiB already!