Enhanced Quick Search, compared it with 2-Way Matching.
The pattern matching code has been greatly enhanced and tuned.

The Quick Search algorithm has been modified (improved) by adding
another computation to the pattern compilation: the period of each
pattern prefix from 0 to the pattern length.  By using the larger
between the delta[] table already computed and the period of the
matched prefix, we can turn down the matching complexity from
O(m*n) to O(n).

The 2-Way String-Matching algorithm from Crochemore-Perrin has been
implemented: it is a known O(n) matching algorithm and is now available
as pattern_match(), an interface similar to pattern_qsearch().

The pattern compilation prepares for both Quick Search or 2-Way, since
the extra running cost is small.  Compilation is done once and is
roughly in O(m).  The extra space requireed to accomodate the two
algorithms is negligible.

Specialized optimized verions of memchr(), strchr() and strlen() are
also implemented.  At startup, pattern_init() will benchmark them
against the libc implementation to choose the best routine.

On linux, the strstr() implementation from the libc is incredibly
efficient, and I cannot seem to be able to write C code that can match
the speed, so this routine must have been carefully crafter in assembly.

To avoid using less efficient matching, we now benchmark the strstr()
routine against our Quick Search, to determine if there is a pattern
length that will make Quick Search faster than strstr().  If there is
not, we will transparently re-route pattern_qsearch() to strstr() at
runtime.  Use pattern_qsearch_force() if the use of Quick Search is
required, regardless of initial benchmarking.

To ensure correctness of the implementation (since I modified Quick
Search), a new pattern-test program performs the necessary regression
tests and can also be used to benchmark the various implementations
with a large or a small alphabet, and with varying pattern lengths!