Optimized allocation of the uperiod[] array in patterns.
Just like the delta[] array, for small patterns we can use 8-bit
quantities, hence the array can be an array of bytes.

The delta[] and uperiod[] arrays, if allocated dynamically, now use
xmalloc() instead of walloc(): since compiled patterns tend to be
compiled once for the duration of the program, it pays to allocate them
in an area that is more for long-term objects.