Improved pattern benchmarking execution time and precision.
The minimal amount of loops and timinig requirements were drastically
reduced, and we compute more data points for statistical analysis.

To accelerate convergence and increase precision of the cut-off point
determination, we use a binary search in the range, with more benchmarking
done when we start to converge to a cut-off point.

The binary search algorithm for cut-off point determination is also run
several times, with increasing precision each time, and with a reduced
alphabet size during refining (which helps exhibit O(mn) behaviours in
non-linear algorithms).

Fixed comment typo.