str_s2c(), str_2c(), str_dup(): handle already NUL-terminated strings.
Since it is possible to explicitly append a NUL character to strings,
there is no need to append an extra NUL character if they already have
a terminating NUL, when the intent is to make them appear as C strings.

Of course, any embedded NUL is ignored and would "truncate" the C string,
but this is not something we want to handle specially.