zalloc: fixed accounting of total amount of blocks.
We now have zn_cram() return the amount of blocks it can cram in a
subzone.  This matters for the first subzone since we can have an
embedded zone structure in there.

Also computes and stores the typical amount of blocks we can cram in
theory for a regular subzone with no embedded zone in it.