Windows: enhanced stack unwinding.
In case we are not correctly figuring out the stack pointer, look at the
return address to parse any CALL instruction preceding it and scan 3
words around the top of the stack to see if any of the addresses there
would not be a better returning point for our unwinding.

The debugging logs of the stack unwinding routine are now done in a
separate file and use a "tracing" format (as opposed to mere logging)
that allows to follow the calls being made during processing.

The log_vbprintf() routine (which detects overflowing in the generated
log string) is now exported, to be reused by other logging routines.