OOB: increased timeout for OOB hit delivery.
With busy ultrapeers doing many OOB-proxying, there can be some delay
between the time we notify them of hits and before they ask us to
deliver the hits.

Hence we increased the timeout they have to reply to our notification,
from 45 to 80 seconds.

Second, the amount of time we keep results in memory is increased for
smaller hits: less hits means either a rare file or a specific query.
This is to make sure we don't discard hits early if it took us time to
(possibly repliably) notify the remote host about the pending hits.