OOB: raise the bar for banning after unclaimed OOB hits.
Threshold was too low and would trigger too soon: it is perfectly legit
for a notified host to not claim hits its OOB for a popular query!

Still, we need to protect against abuse whereby a host would get
queries and store results that would be never claimed, at a "high rate".
That rate is now raised (arbitrarily) to roughly 25 per minute as
opposed to the previous 5 per minute!

Also lowered the maximum ban time to 10 minutes!