XMALLOC: must "embed" xstrdup() and xstrndup().
These routine names must be prefixed with "e_" to embed them, so that
they do not accidentally supersede library symbols expected by the BFD
library!

Internally, we install #define so that our code can still call xstrdup()
and xstrndup() without having to display the additional ugly "e_" prefix.

This probably explains some of the errors I wrongly attributed to bugs
in the BFD library and will be revisiting these assumptions in
subsequent commits!