Added emergency path during startup should sbrk() fail.
This is used when compiling for our own malloc() implementation.

Apparently, sbrk() can fail when running from a Docker environment, so
we define an alternative path involving an early initialization of the
VMM layer, to then attempt to use posix_memalign() instead of sbrk().

I also had to fix some mutex checking logic, as an emergency hack. No
time to investigate why thread_current() and thread_self() diverge on a
freelist mutex, but they do... and this could be another Docker-induced
problem.