Avoid standalone compilation of local_shell.c on Windows.
It would require more than just the C library to be able to link
and run.  Maybe it would on Cygwin but currently Configure considers
Cygwin to be a Windows platform.  If there is a need on Cygwin
platforms, we'll need metaconfig's help to be able to flag Cygwin
differently.