Added upload bandwidth capping.
On follow-up requests, we monitor the apparent transfering rate we're
getting from the previous request and attempt to limit b/w to that
apparent amount, to avoid overwhelming the remote host.

On subsequent follow-ups, depending on how quickly we get the next
request, we can determine how quickly data travel from us to them, and
therefore dynamically adjust up or down the b/w cap.