extract_retry_after(): ensure amount in seconds is the only value.
When parsing a Retry-After header, which can contain a number of seconds
or a date, only treat a plain number (of seconds) as valid when it is the
sole value of the header.

Otherwise, assume a date that we can parse with date2time().

This will avoid wrongly processing headers such as "10:00" as if they
meant 10 seconds!  Not that this last header content would be valid, but
better be safe than sorry.