http_send_status(): ensure generated header has trailing CRLF.
Whenever we add a user-generated HTTP_EXTRA_LINE or invoke an
HTTP_EXTRA_CALLBACK to dynamically generate header information, make
sure the output is properly terminated by a CRLF or loudly warn and
ignore the badly formated material.

Indeed, that would mean we are corrupting the header we are sending
back, so better miss some information than generate a broken reply.

We know the minimal header we construct is properly generated, and we
would only be ignoring extra user-generated information, not critical
to the HTTP protocol.