On HTTP 416, emit a Last-Modified header.
In case their download side is not expecting a Retry-After, maybe they
will parse a Last-Modified and use it to determine a proper retry time
when the requested chunks are not available.

Our own download side now knows how to do that on pure 416 errors, but
gtk-gnutella emits a Retry-After so that processing is only useful when
interacting with other servents.