http_send_status(): reworked to avoid undue header truncation.
When the header cannot be fully generated, perform an emergency close
of the connection to avoid sending back a truncated header that will
only be mis-interpreted by the remote client.

Also introduced the notion of prioritary header callback: when the
header cannot contain all the generated data, we can retry the
generation of a smaller header by keeping only the mandatory (i.e.
prioritary) information generated by these callbacks.  For instance, for
a 206 reply, that would be the Content-Range header!

And if we still have too large a header, even with the limited
mandatory information, then we have no choice but to force a 503 error,
with a "Connection: close" indication, shutdowning the socket to prevent
further exchanges on that connection.