Add lingering period for unreferenced file_descriptor objects.
When file_object_release() removes the last reference on a file
descriptor, instead of immediately removing the associated path from
the file_descriptors table and reclaim the file_descriptor object,
keep the entry for a little while (2 minutes) in case it is going to be
reused soon.

Indeed, creation of the file descriptor involves re-opening the file
it references, which means traversing the file system from the root
directory down to the given file entry.