download_reply(): reload available ranges after 416 if needed.
When we get a 416 "Range not available" and the available ranges we have
for the file (based on the server's advertisements over time) contain
the chunk we requested, it means our list is wrong.  Reload the list of
ranges from scratch, based on the information held in the HTTP reply we
just got.