update_available_ranges(): revisited range processing logic.
Lack of X-Available or X-Available-Ranges headers does not always mean that
the file is complete on the remote server: it depends on the HTTP status
code.  Absence means complete file only when the HTTP status code is 2xx.

Whilst we consolidate advertised ranges over time, it is possible to
get more data consolidated than what the server really holds, when parts
of the ranges it advertised earlier were invalidated somehow remotely.