Ensure stricter memory alignment when xmalloc() replaces malloc().
We need to comply with whatever alignment requirements are imposed
within the native system malloc() in order to be able to safely
substitute our own xmalloc() and use it as the global malloc().

Indeed, it must transparenly work for all the the dynamic libraries
we're loading and which use the malloc() symbol.