Use 429 instead of 550 for "Too Many Requests".
This is the specified code in RFC6585 (April 2012) for HTTP servers to
return when they believe they are under attack.  We extend this to
Gnutella connections as well.

Up to now, GTKG was using 550 as the error code when it chose to reject
an incoming connection due to too frequent attempts.  But this was
chosen way before 2012, at a time where there was no such specification.