Avoid too frequent node connections to given IP:port.
Since we're now going to initiate banning for hosts that attempt to
connect to us too frequently, we need to be careful and avoid hammering
them as well.

We are more stringent on the connection requirement delays because the
other side is only going to see our incoming IP address, hence we cannot
connect too often if several nodes are running locally on our side.