pcache_pong_received(): also validate changing ports.
When a Pong is received from one of our immediate neighbours, we use it
to capture their advertised Gnutella port.

But if we already know their port and it changes over time (apart from
the default GTA_PORT which we initially configure when do not know better
from the handshaking headers), then we flag them as being suspicious
(NODE_F_ALIEN_IP) and emit a warning when debugging.

Also changed node_infostr_to_buf() to display a "~" in front of the
node address if we flagged it with NODE_F_ALIEN_IP.