compat_setjmp(): protect against gone contexts.
Make sure the longjmp() context is still valid, i.e. that the routine
which issued the setjmp() has not already returned when the longjmp()
happens.  As far as we can tell.

Also capture the routine name where the context is taken, in case the
source code changes when we analyse the crash logs!