Made ascii_toupper() and ascii_tolower() avoid branching.
Indeed, avoiding branches helps the instruction pipeline of modern CPUs.

We also tweaked the other is_ascii_xxx() routines to simplify the amount
of instructions required for processing.  The ascii_ctype[] array was
extended to 256 values because the use of '&' in the expressions can
force access to values in the [128-255] range.